requirements.txt
strands-agents>=0.1.1
strands-agents-tools>=0.1.0
exa-py>=1.0.6
streamlit>=1.45.0
python-dotenv>=1.1.0
fastapi[standard]>=0.111.0
uvicorn>=0.29.0
requests>=2.31.0
